refactor(examples): split AutoResearchRemote.cpp into 10 files all under 1K LOC (closes #3132)
Decomposes the 3,960-LOC monolith into a Core shell plus 9 sibling files,
all under 1K LOC. Each new file holds one logical piece of the RPC
machinery:

  - AutoResearchRemote.cpp (Core, 403 LOC): includes + outer helpers
    (printJsonRaw, printStreamRaw), findConnectedPinsImpl, ctor/dtor,
    registerFunctions() shell calling each bind*Methods(),
    tick(), registerAllMethods(), startBleRemote(), stopBleRemote()
  - AutoResearchRemoteRunSingleTest.cpp (750 LOC): anonymous-namespace
    helpers (expectedClocklessWireUs, maxLaneLeds, ScopedFastLedBrightness,
    measureTightTiming) + runSingleTestImpl
  - AutoResearchRemoteRunParallelTest.cpp (379 LOC): runParallelTestImpl
  - AutoResearchRemoteBasicMethods.cpp (157 LOC): status, debugTest,
    deliberateHang, drivers, runSingleTest, runParallelTest, ping,
    testNoSerial
  - AutoResearchRemoteFlexioMethods.cpp (572 LOC): flexioRxBenchmark,
    flexioObjectFledTest, flexioRxLoopbackPing
  - AutoResearchRemoteGpioMethods.cpp (605 LOC): setDebug,
    testGpioConnection, getPins, setTxPin, setRxPin, setPins,
    findConnectedPins, help
  - AutoResearchRemoteBenchmarkMethods.cpp (409 LOC): testSimd,
    testSimdBenchmark, animartrixPerlinBench, wave8ExpandBenchmark,
    parlioStreamValidate, parlioEncodeBenchmark
  - AutoResearchRemoteAsyncNetMethods.cpp (369 LOC): testAsync,
    startNetServer, startNetClient, runNetClientTest, runNetLoopback,
    stopNet, startOta, stopOta, startBle, stopBle, decodeFile, bleStatus
  - AutoResearchRemoteCoroutineMethods.cpp (504 LOC): testCoroutineBasic,
    Stop, Concurrent, Await, AwaitError, PromiseCallbacks,
    PromiseCatchCallback, ChainedAwait, All
  - AutoResearchRemotePerfMethods.cpp (286 LOC): perfProbeMemcpy,
    perfProbeNop, perfProbeRepeat, wave2dPerf

Split mechanism: each bind*Methods() is a private member of
AutoResearchRemoteControl, declared in AutoResearchRemote.h. Each
function body lives in its own .cpp file. The lambdas inside each bind
use [this] capture and access private members (mRemote, mState,
runSingleTestImpl, etc.) naturally — same access rules as before the
split. No public-API change; only file partitioning.

registerFunctions() is now a 7-line shell that calls each bind*Methods()
in order. RPC registration order is preserved.

Build verification: bash compile wasm --examples AutoResearch — SUCCESS
(library 34.84s, sketch 16.40s, link 3.52s, total 81.87s).

Closes #3132. Meta tracking issue: #3127.
